+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#1507](https://github.com/alleslabs/celatone-frontend/pull/1507) Fix balance change redirection
+
 ## v1.16.0
 
 ### Features